backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.7
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core
langchain-text-splitters
langsmith>=0.3.45
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.3
nodeenv==1.9.1
numpy==1.26.4
openai>=1.35.13,<2.0.0
orjson==3.10.18
packaging==24.2
prisma==0.13.0
propcache==0.3.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.19.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.8
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.37.2
tenacity==8.5.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.0
typing_extensions==4.14.0
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.30.1
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
